requirements.txt
# AI Job Displacement Analysis - Project Dependencies
# Python 3.9+
# Install with: pip install -r requirements.txt

# Data Processing & Analysis
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1

# Machine Learning & Modeling
scikit-learn==1.2.2
xgboost==1.7.5
lightgbm==4.0.0

# Model Interpretation & Explainability
shap==0.42.1
eli5==0.13.2

# Visualization & Plotting
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.13.0
altair==4.2.2

# Jupyter & Notebooks
jupyter==1.0.0
jupyterlab==3.6.3
ipykernel==6.22.0
ipywidgets==8.0.6
notebook==6.5.3

# Code Quality & Testing
pytest==7.3.1
pytest-cov==4.1.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.2.0

# Development Tools
ipython==8.12.2
python-dotenv==1.0.0
pyyaml==6.0
tqdm==4.65.0

# Web Framework (for deployment)
flask==2.3.1
streamlit==1.22.1
gunicorn==20.1.0

# Data I/O Formats
pyarrow==12.0.0
openpyxl==3.1.2

# Utilities
requests==2.31.0
click==8.1.3
pydantic==1.10.8
